python: add exception types
